// src/app/model/month-hours.ts
import { Resource } from './resource';

export class MonthHours extends Resource {
  year: string;
  january: number;
  february: number;
  march: number;
  april: number;
  may: number;
  june: number;
  july: number;
  august: number;
  september: number;
  october: number;
  november: number;
  december: number;
}